+++ tests/testthat.R
@@ -1,4 +1,5 @@
 options(Matrix.warnDeprecatedCoerce = 2)
+options(lifecycle_verbosity = "quiet")
 
 library(testthat)
 library(sdmTMB)